Resolve and canonicalize URLs from crawl requests

+++ reciperadar/workers/recipes.py
@@ -1,5 +1,5 @@
 from reciperadar.models.recipes import Recipe
-from reciperadar.models.url import RecipeURL
+from reciperadar.models.url import CrawlURL, RecipeURL
 from reciperadar.services.database import Database
 from reciperadar.workers.broker import celery
 
@@ -39,7 +39,7 @@ def crawl_recipe(url):
     recipe_url = session.query(RecipeURL).get(url) or RecipeURL(url=url)
 
     try:
-        recipe_json = recipe_url.crawl()
+        response = recipe_url.crawl()
     except RecipeURL.BackoffException:
         print(f'Backoff: {recipe_url.error_message} for url={recipe_url.url}')
         return
@@ -51,11 +51,45 @@ def crawl_recipe(url):
         session.commit()
         session.close()
 
-    recipe = Recipe.from_doc(recipe_json)
+    if not response.ok:
+        return
+
+    recipe = Recipe.from_doc(response.json())
     process_recipe.delay(recipe.id)
 
     session = Database().get_session()
     session.query(Recipe).filter_by(id=recipe.id).delete()
     session.add(recipe)
+    session.commit()
+    session.close()
+
+
+@celery.task(queue='crawl_url')
+def crawl_url(url):
+    session = Database().get_session()
+    crawl_url = session.query(CrawlURL).get(url) or CrawlURL(url=url)
+
+    try:
+        response = crawl_url.crawl()
+    except RecipeURL.BackoffException:
+        print(f'Backoff: {crawl_url.error_message} for url={crawl_url.url}')
+        return
+    except Exception:
+        print(f'{crawl_url.error_message} for url={crawl_url.url}')
+        return
+    finally:
+        session.add(crawl_url)
+        session.commit()
+        session.close()
+
+    if not response.ok:
+        return
+
+    recipe_url = RecipeURL(url=response.url)
+    crawl_recipe.delay(recipe_url.url)
+
+    session = Database().get_session()
+    session.query(RecipeURL).filter_by(url=recipe_url.url).delete()
+    session.add(recipe_url)
     session.commit()
     session.close()
